requirements.txt
scipy
tenacity
tiktoken==0.3.3
termcolor 
openai
requests
arxiv
pandas
PyPDF2
tqdm
python-dotenv
gradio